chore: refactored marquee files for better organization
