+++ src/components/Marquee/index.tsx
